update twist derivative part in twist2npy

+++ qolo/twist2npy.py
@@ -19,6 +19,7 @@ import numpy as np
 import rosbag
 
 from crowdbot_data import AllFrames, bag_file_filter
+from interp_util import interp_translation
 
 #%% extract twist_stamped from rosbag without rosbag play
 def timestamp_str(ts):
@@ -54,14 +55,46 @@ def extract_twist_from_rosbag(bag_file_path, args):
                 twist_msg_sum += 1
 
     t_np = np.asarray(t_list, dtype=np.float64)
-    x_np = np.asarray(x_list, dtype=np.float32)
-    z_np = np.asarray(zrot_list, dtype=np.float32)
+    x_np = np.asarray(x_list, dtype=np.float64)
+    z_np = np.asarray(zrot_list, dtype=np.float64)
 
     twist_stamped_dict = {'timestamp': t_np, 
                           'x': x_np, 
                           'zrot': z_np}
     print('Current twist extracted!')
     return twist_stamped_dict
+
+def interp_twist(twist_stamped_dict, target_dict):
+    """Calculate interpolations for all states
+    """
+    source_ts = twist_stamped_dict.get('timestamp')
+    source_x = twist_stamped_dict.get('x')
+    source_zrot = twist_stamped_dict.get('zrot')
+    interp_ts = target_dict.get('timestamp')
+    interp_ts = np.asarray(interp_ts, dtype=np.float64)
+    # don't resize, just discard the timestamps smaller than source
+    if min(interp_ts) < min(source_ts):
+        interp_ts[interp_ts<min(source_ts)] = min(source_ts)
+    elif max(interp_ts) > max(source_ts):
+        interp_ts[interp_ts>max(source_ts)] = max(source_ts)
+    
+    interp_dict = {}
+    interp_dict['timestamp'] = interp_ts
+    interp_dict['x'] = interp_translation(source_ts, interp_ts, source_x)
+    interp_dict['zrot'] = interp_translation(source_ts, interp_ts, source_zrot)
+    return interp_dict
+
+def compute_twist_derivative(twist_stamped_dict):
+    source_ts = twist_stamped_dict.get('timestamp')
+    source_x = twist_stamped_dict.get('x')
+    source_zrot = twist_stamped_dict.get('zrot')
+    x_acc = np.gradient(source_x, source_ts, axis=0)
+    zrot_acc = np.gradient(source_zrot, source_ts, axis=0)
+    acc_stamped_dict = {'timestamp': source_ts, 
+                          'x': x_acc, 
+                          'zrot': zrot_acc}
+    print('Current twist extracted!')
+    return acc_stamped_dict
 
 #%% main file
 if __name__ == "__main__":
@@ -75,9 +108,12 @@ if __name__ == "__main__":
                         help='different subfolder in rosbag/ dir')
     parser.add_argument('--twist_topic', default='/qolo/twist', type=str,
                         help='topic for qolo twist')
+    parser.add_argument('--gen_acc', dest='gen_acc', action='store_true',
+                        help="Whether to generate the acceleration (default: true)")
+    parser.set_defaults(gen_acc=True)
     parser.add_argument('--overwrite', dest='overwrite', action='store_true',
                         help="Whether to overwrite existing rosbags (default: false)")
-    parser.set_defaults(feature=False)
+    parser.set_defaults(overwrite=False)
     args = parser.parse_args()
 
     allf = AllFrames(args)
@@ -86,9 +122,10 @@ if __name__ == "__main__":
     rosbag_dir = os.path.join(args.base, args.data, "rosbag", args.folder)
     bag_files = list(filter(bag_file_filter, os.listdir(rosbag_dir))) 
     
-    # destination: twist data in data/xxxx_processed/twist
-    if not os.path.exists(allf.twist_dir):
-        os.makedirs(allf.twist_dir)
+    # destination: twist data in data/xxxx_processed/source_data/twist
+    twist_dir = os.path.join(allf.source_data_dir, 'twist')
+    if not os.path.exists(twist_dir):
+        os.makedirs(twist_dir)
 
     print("Starting extracting twist_stamped from {} rosbags!".format(len(bag_files)))
 
@@ -99,18 +136,49 @@ if __name__ == "__main__":
         counter += 1
         print("({}/{}): {}".format(counter, len(bag_files), bag_path))
 
-        twist_stamped_filepath = os.path.join(allf.twist_dir, bag_name+'_twist_stamped.npy')
+        twist_stamped_filepath = os.path.join(twist_dir, bag_name+'_twist_stamped.npy')
+        # sample with lidar frame
+        twist_sampled_filepath = os.path.join(twist_dir, bag_name+'_twist_sampled.npy')
 
-        if (not os.path.exists(twist_stamped_filepath)) or (args.overwrite):
-            twist_stamped_dict = extract_twist_from_rosbag(bag_path, args)
-            np.save(twist_stamped_filepath, twist_stamped_dict)
-        else:
-            print("Detecting the generated {} already existed!".format(twist_stamped_filepath))
-            print('Will not overwrite. If you want to overwrite, use flag --overwrite')
-            continue
+        if not os.path.exists(twist_sampled_filepath):
+            if (not os.path.exists(twist_stamped_filepath)) or (args.overwrite):
+                twist_stamped_dict = extract_twist_from_rosbag(bag_path, args)
+                np.save(twist_stamped_filepath, twist_stamped_dict)
+            else:
+                print("Detecting the generated {} already existed!".format(twist_stamped_filepath))
+                print('If you want to overwrite, use flag --overwrite')
+                twist_stamped_dict = np.load(twist_stamped_filepath, allow_pickle=True).item()
+
+            lidar_stamped = np.load(os.path.join(allf.lidar_dir, bag_name+"_stamped.npy"), allow_pickle=True).item()
+            twist_sampled_dict = interp_twist(twist_stamped_dict, lidar_stamped)
+            np.save(twist_sampled_filepath, twist_sampled_dict)
+            
+            source_len = len(twist_stamped_dict['timestamp'])
+            target_len = len(lidar_stamped['timestamp'])
+            print("# Sample from {} frames to {} frames.".format(source_len, target_len))
+
+        if args.gen_acc:
+            acc_dir = os.path.join(allf.source_data_dir, 'acc')
+            if not os.path.exists(acc_dir):
+                os.makedirs(acc_dir)
+            acc_stamped_filepath = os.path.join(acc_dir, bag_name+'_acc.npy')
+            # sample with lidar frame
+            acc_sampled_filepath = os.path.join(acc_dir, bag_name+'_acc_sampled.npy')
+
+            if (not os.path.exists(acc_stamped_filepath)) or (args.overwrite):
+                if not ('twist_stamped_dict' in locals().keys()):
+                    twist_stamped_dict = np.load(twist_stamped_filepath, allow_pickle=True).item()
+                acc_stamped_dict = compute_twist_derivative(twist_stamped_dict)
+                np.save(acc_stamped_filepath, acc_stamped_dict)
+
+                if not ('lidar_stamped' in locals().keys()):
+                    lidar_stamped = np.load(os.path.join(allf.lidar_dir, bag_name+"_stamped.npy"), allow_pickle=True).item()
+                acc_sampled_dict = interp_twist(acc_stamped_dict, lidar_stamped)
+                np.save(acc_sampled_filepath, acc_sampled_dict)
     
-    print("Finish extracting all twist_stamped msg!")
-
+    print("Finish extracting all twist!")
+    if args.gen_acc:
+        print("& computing acc (derivative of twist)!")
 """
 ref: https://github.com/uzh-rpg/rpg_e2vid/blob/master/scripts/extract_events_from_rosbag.py
 http://wiki.ros.org/rosbag/Code%20API#Python_API
